Refactor the splice funding-feerate helpers into fee_estimator.rs
The 1.5x-of-estimate funding feerate ceiling was open-coded identically
in splice_in and splice_out. Route both through a max_funding_feerate
helper and keep it, alongside rbf_splice_feerates, in fee_estimator.rs
so the splice funding-feerate policy lives in one place.
